+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state.dart
@@ -198,6 +198,10 @@ abstract class MPTH2FileEditState {
     return Future.value();
   }
 
+  Future<void> onPrimaryButtonDoubleClick(PointerUpEvent event) {
+    return Future.value();
+  }
+
   void onSecondaryButtonClick(PointerUpEvent event) {}
 
   void onTertiaryButtonClick(PointerUpEvent event) {}