+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_empty_selection.dart
@@ -123,6 +123,49 @@ class MPTH2FileEditStateSelectEmptySelection extends MPTH2FileEditState
     );
   }
 
+  @override
+  Future<void> onPrimaryButtonDoubleClick(PointerUpEvent event) async {
+    final Map<int, THElement> clickedLineSegments = selectionController
+        .getSelectableElementsClickedWithoutDialog(
+          screenCoordinates: event.localPosition,
+          selectionType: THSelectionType.lineSegment,
+        );
+
+    if (clickedLineSegments.isNotEmpty) {
+      final THLineSegment clickedLineSegment =
+          clickedLineSegments.values.first as THLineSegment;
+      final THLine parentLine = th2File.lineByMPID(
+        clickedLineSegment.parentMPID,
+      );
+
+      selectionController.setSelectedElements(<THElement>[parentLine]);
+      th2FileEditController.stateController.setState(
+        MPTH2FileEditStateType.editSingleLine,
+      );
+
+      return;
+    }
+
+    final Map<int, THElement> clickedElements = selectionController
+        .getSelectableElementsClickedWithoutDialog(
+          screenCoordinates: event.localPosition,
+          selectionType: THSelectionType.pla,
+        );
+
+    for (final THElement clickedElement in clickedElements.values) {
+      if (clickedElement is! THLine) {
+        continue;
+      }
+
+      selectionController.setSelectedElements(<THElement>[clickedElement]);
+      th2FileEditController.stateController.setState(
+        MPTH2FileEditStateType.editSingleLine,
+      );
+
+      return;
+    }
+  }
+
   /// Draws the selection window.
   @override
   void onPrimaryButtonDragUpdate(PointerMoveEvent event) {